requirements.txt
aniso8601==9.0.1
azure-core==1.15.0
azure-storage-blob==12.8.1
beautifulsoup4==4.9.3
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
cycler==0.10.0
Flask==2.0.1
Flask-RESTful==0.3.9
google==3.0.0
google-api-core==1.30.0
google-auth==1.31.0
google-cloud==0.34.0
google-cloud-core==1.7.0
google-cloud-storage==1.38.0
google-crc32c==1.1.2
google-resumable-media==1.3.0
googleapis-common-protos==1.53.0
idna==2.10
isodate==0.6.0
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
msrest==0.6.21
numpy==1.20.3
oauthlib==3.1.1
opencv-python==4.5.2.54
packaging==20.9
Pillow==8.2.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.6.3
six==1.16.0
soupsieve==2.2.1
urllib3==1.26.5
Werkzeug==2.0.1
